+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased 3.0](https://github.com/opensearch-project/k-NN/compare/2.x...HEAD)
 ### Features
+* Added a boolean flag that enable user to use Lucene searcher on FAISS index. [#2405](https://github.com/opensearch-project/k-NN/pull/2405)
 ### Enhancements
 ### Bug Fixes
 ### Infrastructure